pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "GenomeUtils"
version = "0.1.1"
description = "An object-oriented Python library for working with genomic data."
authors = [
    {name="Seyedarash Ayatollahi", email="seyedarash.ayatollahi@b-tu.de"},
]
maintainers = [
    {name="Nathalie Gocht", email="nathalie.gocht@b-tu.de"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent"
]
license = "LGPL-3.0-or-later"

dependencies = [
    "gffutils>=0.13,<0.14",
    "requests>=2.32,<3",
    "biopython>=1.83,<2",
    "gget>=0.29,<0.30",
    "tqdm>=4.67,<5",
]

[tool.setuptools.packages.find]
include = ["GenomeUtils*"]
